chore: Extract scope manager concurrent continuation
